6-size.c

+++ 0x00-hello_world/6-size.c
@@ -10,10 +10,10 @@ int main(void)
 	long int c;
 	long long int d;
 	float f;
-printf("size of char:%zu byte(s)\n", sizeof(a));
-printf("size of int:%zu byte(s)\n", sizeof(b));
-printf("size of long int :%zu byte(s)\n", sizeof(c));
-printf("size of long long int:%zu byte(s)\n", sizeof(d));
-printf("size of float:%zu byte(s)\n", sizeof(f));
-return 0 ;
+printf("size of char:%zu byte(s)\n",(unsigned long)sizeof(a));
+printf("size of int:%zu byte(s)\n", (unsigned long)sizeof(b));
+printf("size of long int :%zu byte(s)\n", (unsigned long)sizeof(c));
+printf("size of long long int:%zu byte(s)\n", (unsigned long)sizeof(d));
+printf("size of float:%zu byte(s)\n", (unsigned long)sizeof(f));
+return (0);
 }
